+++ Xabbo.Core/Entity/IPet.cs
@@ -5,7 +5,7 @@ namespace Xabbo.Core
     public interface IPet : IEntity
     {
         int Breed { get; }
-        int OwnerId { get; }
+        long OwnerId { get; }
         string OwnerName { get; }
         int RarityLevel { get; }
         int Level { get; }
