+++ Xabbo.Core/Entity/IRoomUser.cs
@@ -5,7 +5,7 @@ namespace Xabbo.Core
     public interface IRoomUser : IEntity
     {
         Gender Gender { get; }
-        int GroupId { get; }
+        long GroupId { get; }
         string GroupName { get; }
         string FigureExtra { get; }
         int AchievementScore { get; }
